+++ ldes-server-compaction/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/compaction/application/ViewDeletedHandlerCompaction.java
@@ -1,6 +1,6 @@
 package be.vlaanderen.informatievlaanderen.ldes.server.compaction.application;
 
-import be.vlaanderen.informatievlaanderen.ldes.server.compaction.domain.ViewCollection;
+import be.vlaanderen.informatievlaanderen.ldes.server.compaction.domain.repository.ViewCollection;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.view.valueobject.ViewDeletedEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
