+++ ldes-server-compaction/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/compaction/application/events/FragmentsCompactedEvent.java
@@ -1,4 +1,4 @@
-package be.vlaanderen.informatievlaanderen.ldes.server.compaction.domain.services;
+package be.vlaanderen.informatievlaanderen.ldes.server.compaction.application.events;
 
 import be.vlaanderen.informatievlaanderen.ldes.server.fragmentation.entities.Fragment;
